refactor(bridge): rename framework entrypoints to dedicated basenames
Renames Angular, Lit, and React entrypoints via git mv to preserve
git history.

This makes integration into google source simpler.
