+++ bridge/src/angular/angular-bridge.ts
