+++ bridge/src/lit/lit-bridge.spec.ts
@@ -16,7 +16,7 @@
 
 // @vitest-environment jsdom
 import {describe, it, expect, afterEach, vi} from 'vitest';
-import {bootstrapLitSandbox, A2uiSandboxRoot} from './index';
+import {bootstrapLitSandbox, A2uiSandboxRoot} from './lit-bridge';
 import {a2uiBridge} from '../preview-bridge';
 import {Catalog, ComponentApi} from '@a2ui/web_core/v0_9';
